Split functions in genmap.c to smaller files
