+++ src/genmap.c
@@ -1,4 +1,3 @@
-#define _POSIX_C_SOURCE 200112
 #include "genmap-impl.h"
 #include "genmap-io.h"
 
@@ -91,107 +90,6 @@ int GenmapFinalize(GenmapHandle h) {
   GenmapDestroyHandle(h);
 
   return 0;
-}
-//
-// GenmapHandle
-//
-int GenmapCreateHandle(GenmapHandle h) {
-  // Datastructures
-  h->global = NULL;
-  h->local  = NULL;
-
-  h->elementArray.ptr = NULL;
-  h->elementArray.n = h->elementArray.max = 0;
-
-  h->Create(h);
-
-  return 0;
-}
-
-int GenmapDestroyHandle(GenmapHandle h) {
-  GenmapFree(h);
-  return 0;
-}
-//
-// GenmapComm
-//
-int GenmapCreateComm(GenmapComm *c, GenmapCommExternal ce) {
-  GenmapMalloc(1, c);
-  comm_init(&(*c)->gsComm, ce);
-  (*c)->verticesHandle = NULL;
-  (*c)->laplacianWeights = NULL;
-  buffer_init(&(*c)->buf, 1024);
-  return 0;
-}
-
-int GenmapDestroyComm(GenmapComm c) {
-  if(&c->buf)
-    buffer_free(&c->buf);
-  if(c->verticesHandle)
-    gs_free(c->verticesHandle);
-  if(c->laplacianWeights)
-    GenmapFree(c->laplacianWeights);
-  if(&c->gsComm)
-    comm_free(&c->gsComm);
-  GenmapFree(c);
-
-  return 0;
-}
-
-int GenmapCommSize(GenmapComm c) {
-  return (int) c->gsComm.np;
-}
-
-int GenmapCommRank(GenmapComm c) {
-  return (int) c->gsComm.id;
-}
-
-GenmapElements GenmapGetElements(GenmapHandle h) {
-  return (GenmapElements) h->elementArray.ptr;
-}
-
-void GenmapSetElements(GenmapHandle h, GenmapElements elements) {
-  h->elementArray.ptr = elements;
-}
-
-GenmapComm GenmapGetLocalComm(GenmapHandle h) {
-  return h->local;
-}
-
-void GenmapSetLocalComm(GenmapHandle h, GenmapComm c) {
-  h->local = c;
-}
-
-GenmapComm GenmapGetGlobalComm(GenmapHandle h) {
-  return h->global;
-}
-
-void GenmapSetGlobalComm(GenmapHandle h, GenmapComm c) {
-  h->global = c;
-}
-
-GenmapInt GenmapGetNLocalElements(GenmapHandle h) {
-  return h->lelt;
-}
-
-void GenmapSetNLocalElements(GenmapHandle h, GenmapInt localElements) {
-  h->lelt = localElements;
-}
-
-GenmapLong GenmapGetNGlobalElements(GenmapHandle h) {
-  return h->nel;
-}
-
-void GenmapSetNGlobalElements(GenmapHandle h, GenmapLong globalElements) {
-  h->nel = globalElements;
-}
-
-GenmapLong GenmapGetLocalStartIndex(GenmapHandle h) {
-  return h->start;
-}
-
-void GenmapSetLocalStartIndex(GenmapHandle h, GenmapLong localStart) {
-  h->start = localStart;
 }
 //
 // GenmapMalloc, Realloc, Calloc and Free
